requirements.txt
requests==2.32.3
pytest==8.4.1
allure-pytest==2.14.3
selenium==4.1.0